MessageType persistence model
I've created an abstract class named MessageType that will be used as a parent to subclasses that will then implement the necessary methods.

This abstract class is mapped to a domain table in the DB, where I used Hibernate annotations in order to configure these mappings.

Each subclass represents a different way to send a scheduled message, and to instantiate these sub-classes, I've created a Simple Factory.

These subclasses already have their own implementation of the method responsible to validate if the recipient is valid for that type of message (e.g. the Email type validates if the recipient is a valid email, whereas the SMS type validates if the recipient is a valid cell phone number, and so on).