ScheduleMessage persistence model
Created the persistence model for the ScheduledMessage class.

And just like the MessageType class, I've used Hibernate annotations to configure its mappings to the DB entities.

I've also created the factory that will instantiate this class based on the MessageType that is passed as a parameter.

The ScheduledMessage class is already responsible to validate its own data to help to ensure that nothing wrong will be saved on the DB.