Adding the DTOs models
In order to avoid sending the persistence models to the client, I've decided to use DTO models.

I've created a DTO model for the ScheduledMessage class, another for the MessageTYpe class, and a whole new DTO that will later be used by the Controller as a generic form to send a standard response to the client.

In addition to that, by using the MapStruct library, I've setup a code that will be responsible to "convert" a DTO instance into a persistence model and another one to do the opposite of that, "convert" a persistence model object into a DTO.