Improvement on the MessageType handling
The MessageType factory now uses the Strategy pattern. By using this pattern we can remove the need to change our factory whenever we need to make a new implementation of the MessageType abstract class.

I've also removed the ScheduleMessage related factories as I wasn't using them, and also because the way I've implemented it didn't make too much sense.

The MessageType abstract class now performs a default validation of the recipient assuming that it needs to be a valid cell phone number.

The ScheduleMessageController no longer ensures whether or not the Message Type sent in the request exists in the DB (the Service is still doing that, though). This Controller now validates if the MessageType is present on the Enum.