Added some Unit Tests
Added a few Unit Tests to the project.

As I had tons of problems to make Maven recognize my tests, I`ve only had time to write a few of them.
Had to make some changes in the pom.xml in order to achieve that.

I`ve added Jacoco as a dependency so I could evaluate how much of my code is covered by testing (which right now is far less that I would like).